- Fixed non-interlace modes to use ClearImage instead of ClearImage2 - Removed nonsense uint64, replaced it with a proper 64bit one (won't be using that much) - Added Gfx_AddPrims function to submit prims in a batch - Added PRIM_FLAG_PERSP to specify that the primitives will need perspective transformation. If it's not set, primitives will be assumed to be already in screen space.

+++ projects/sandbox/source/engine/core/core.c
@@ -1,10 +1,10 @@
 #include "core.h"
 
-uint64 __ramsize    = 0x00200000; // force 2MB of RAM
-uint64 __stacksize  = 0x00004000; // force 16KB of stack
+uint32 __ramsize    = 0x00200000; // force 2MB of RAM
+uint32 __stacksize  = 0x00004000; // force 16KB of stack
 
-uint64 g_ram_start = 0x8001d800;
-uint64 g_ram_end = 0x801f8000;
+uint32 g_ram_start = 0x8001d800;
+uint32 g_ram_end = 0x801f8000;
 
 ///////////////////////////////////////////////////
 int16 Core_Initialize()
