+++ projects/sandbox/source/engine/engine.h
@@ -10,15 +10,17 @@
 #include <libpress.h>
 #include <libcd.h>
 
+#include <limits.h>
+
 // Data types
-typedef unsigned char   uint8;
-typedef unsigned short  uint16;
-typedef unsigned int    uint32;
-typedef unsigned long   uint64;
-typedef char            int8;
-typedef short           int16;
-typedef int             int32;
-typedef long            int64;
+typedef unsigned char			uint8;
+typedef unsigned short			uint16;
+typedef unsigned int			uint32;
+typedef unsigned long long		uint64;
+typedef char					int8;
+typedef short					int16;
+typedef int						int32;
+typedef long long				int64;
 
 #ifndef NULL
 #define NULL (0)
